• Mise en place et test de l'interpolation et des "step/sec , speed" terminés Erreur : En fonction de la position de la plateforme le "RESET" n'est pas toujours effectué et après un "RESET" la plateforme ne se déplace plus • Correction de l'erreur => pour le l'ordre, la position de départ changeait pendant les opérations de déplacement de la plateforme. Pour le bouton, le timer et le port Serial devaient être fermés avant le reset et rouvert après. • Mise en place de la sauvegarde du code en fichier ".sim" et de l'ouverture de fichier ".sim" • Les morceaux de code avec indentation et commentaire (avec des #) sont maintenant pris en charge par le programme et ne génère plus d'erreurs • Ajout d'un message de notification lorsque, lors du lancement d'un déplacement, la zone de travail va être franchie • Modification du message de notification. Ajout d'une demande pour stopper le programme (Yes/No) si oui le programme est reset sinon le déplacement va être effectué mais sans franchir la zone de travail de la plateforme

+++ Variable.cs
@@ -10,8 +10,8 @@ namespace language_prog_simu_6DOF
     {
         public int id { get; set; }
         public string name { get; set; }
-        public int val { get; set; }
-        public Variable(int id, string name, int val)
+        public double val { get; set; }
+        public Variable(int id, string name, double val)
         {
             this.id = id;
             this.name = name;
